refactor: PlaylistDetailの副作用をカスタムフックに切り出して責務を分離
